requirements.txt
auth0-python==2.0.1